manual-tools/requirements.txt
fastapi>=0.110,<0.115
pydantic>=2.6,<3
uvicorn[standard]>=0.29,<0.33
PyYAML>=6.0,<7
pytest>=7,<9
httpx>=0.24,<1